index refactoring + docs
- add a lot of depreciations
- moved some index related files
- docs: add query exemples
