+++ src/test/java/index/SIndexBench.java
@@ -1,4 +1,4 @@
-package db.index;
+package index;
 
 import java.io.FileInputStream;
 import java.io.IOException;
