+++ src/test/java/index/package-info.java
@@ -6,4 +6,4 @@
  * 
  *
  */
-package db.index;
+package index;
